cleanup: separate out service-worker form building logic
move service_worker specific metadata to service_worker.rs
