--- src/settings/metadata.rs
@@ -1,9 +1,0 @@
-use serde::Serialize;
-
-use crate::settings::binding::Binding;
-
-#[derive(Serialize, Debug)]
-pub struct Metadata {
-    pub body_part: String,
-    pub bindings: Vec<Binding>,
-}
