+++ src/settings/mod.rs
@@ -2,7 +2,6 @@ pub mod binding;
 mod environment;
 mod global_config;
 pub mod global_user;
-pub mod metadata;
 pub mod toml;
 
 pub use environment::{Environment, QueryEnvironment};
