+++ src/upload/form/mod.rs
@@ -1,21 +1,21 @@
 mod plain_text;
 mod project_assets;
+mod service_worker;
 mod text_blob;
 mod wasm_module;
 
-use reqwest::blocking::multipart::{Form, Part};
+use reqwest::blocking::multipart::Form;
 use std::fs;
 use std::path::Path;
 use std::path::PathBuf;
 
 use crate::settings::binding;
-use crate::settings::metadata::Metadata;
 use crate::settings::toml::{Target, TargetType};
 use crate::sites::AssetManifest;
 use crate::wranglerjs;
 
 use plain_text::PlainText;
-use project_assets::ProjectAssets;
+use project_assets::ServiceWorkerAssets;
 use text_blob::TextBlob;
 use wasm_module::WasmModule;
 
@@ -60,7 +60,7 @@ pub fn build(
             wasm_modules.push(wasm_module);
             let script_path = PathBuf::from("./worker/generated/script.js");
 
-            let assets = ProjectAssets::new(
+            let assets = ServiceWorkerAssets::new(
                 script_path,
                 wasm_modules,
                 kv_namespaces.to_vec(),
@@ -68,7 +68,7 @@ pub fn build(
                 plain_texts,
             )?;
 
-            build_form(&assets, session_config)
+            service_worker::build_form(&assets, session_config)
         }
         TargetType::JavaScript => {
             log::info!("JavaScript project detected. Publishing...");
@@ -77,7 +77,7 @@ pub fn build(
 
             let script_path = package.main(&package_dir)?;
 
-            let assets = ProjectAssets::new(
+            let assets = ServiceWorkerAssets::new(
                 script_path,
                 wasm_modules,
                 kv_namespaces.to_vec(),
@@ -85,7 +85,7 @@ pub fn build(
                 plain_texts,
             )?;
 
-            build_form(&assets, session_config)
+            service_worker::build_form(&assets, session_config)
         }
         TargetType::Webpack => {
             log::info!("webpack project detected. Publishing...");
@@ -110,7 +110,7 @@ pub fn build(
                 text_blobs.push(text_blob);
             }
 
-            let assets = ProjectAssets::new(
+            let assets = ServiceWorkerAssets::new(
                 script_path,
                 wasm_modules,
                 kv_namespaces.to_vec(),
@@ -118,7 +118,7 @@ pub fn build(
                 plain_texts,
             )?;
 
-            build_form(&assets, session_config)
+            service_worker::build_form(&assets, session_config)
         }
     }
 }
@@ -126,72 +126,6 @@ pub fn build(
 fn get_asset_manifest_blob(asset_manifest: AssetManifest) -> Result<String, failure::Error> {
     let asset_manifest = serde_json::to_string(&asset_manifest)?;
     Ok(asset_manifest)
-}
-
-fn build_form(
-    assets: &ProjectAssets,
-    session_config: Option<serde_json::Value>,
-) -> Result<Form, failure::Error> {
-    let mut form = Form::new();
-
-    // The preview service in particular streams the request form, and requires that the
-    // "metadata" part be set first, so this order is important.
-    form = add_metadata(form, assets)?;
-    form = add_files(form, assets)?;
-    if let Some(session_config) = session_config {
-        form = add_session_config(form, session_config)?
-    }
-
-    log::info!("building form");
-    log::info!("{:?}", &form);
-
-    Ok(form)
-}
-
-fn add_files(mut form: Form, assets: &ProjectAssets) -> Result<Form, failure::Error> {
-    form = form.file(assets.script_name(), assets.script_path())?;
-
-    for wasm_module in &assets.wasm_modules {
-        form = form.file(wasm_module.filename(), wasm_module.path())?;
-    }
-
-    for text_blob in &assets.text_blobs {
-        let part = Part::text(text_blob.data.clone())
-            .file_name(text_blob.binding.clone())
-            .mime_str("text/plain")?;
-
-        form = form.part(text_blob.binding.clone(), part);
-    }
-
-    Ok(form)
-}
-
-fn add_metadata(mut form: Form, assets: &ProjectAssets) -> Result<Form, failure::Error> {
-    let metadata_json = serde_json::json!(&Metadata {
-        body_part: assets.script_name(),
-        bindings: assets.bindings(),
-    });
-
-    let metadata = Part::text((metadata_json).to_string())
-        .file_name("metadata.json")
-        .mime_str("application/json")?;
-
-    form = form.part("metadata", metadata);
-
-    Ok(form)
-}
-
-fn add_session_config(
-    mut form: Form,
-    session_config: serde_json::Value,
-) -> Result<Form, failure::Error> {
-    let wrangler_session_config = Part::text((session_config).to_string())
-        .file_name("")
-        .mime_str("application/json")?;
-
-    form = form.part("wrangler-session-config", wrangler_session_config);
-
-    Ok(form)
 }
 
 fn filename_from_path(path: &PathBuf) -> Option<String> {
